+++ fs/cnmt.js
@@ -4,7 +4,7 @@ const HEXES = new Array(256).fill().map((_, i) => i.toString(16).padStart(2, '0'
 // Values verified against hacPack cnmt.c (cnmt_create_application: type = 0x1 Program /
 // 0x2 Data / 0x3 Control / 0x4 HtmlDocument / 0x5 LegalInformation) and
 // sinjunyoung/lh Ncm/ContentEnums.cs. Source: https://github.com/DarkMatterCore/hacPack/blob/master/cnmt.c
-export const CONTENT_TYPE = {
+export const CNMT_ENTRY_TYPE = {
     META: 0,
     PROGRAM: 1,
     DATA: 2,